feat: add duel announcements

+++ migrations/0005_create-duel-participants-table.sql
@@ -1,7 +1,7 @@
 -- Up Migration
 CREATE TABLE duel_participants (
   duel_id INT NOT NULL REFERENCES duels (id) ON DELETE CASCADE,
-  user_id INT NOT NULL REFERENCES users (id) ON DELETE CASCADE,
+  user_id BIGINT NOT NULL REFERENCES users (id) ON DELETE CASCADE,
   PRIMARY KEY (duel_id, user_id)
 );
